Fix lookup of lib[n]curses.
This commit revises a bit how a working curses implementation is identified
and how it is checked that readline links fine with that.
First, look for ncurses. If that is not found, look for curses. Then check
whether readline works with what we found. Additionally, this fixes a couple
of mostly innocent problems:

* -l[n]curses was erroneously added to CPPFLAGS
* if found, both -lcurses and -lncurses were added to LIBS