Fix get1char() to also use std::cin like the new getin() does.
After these changes "make check" passes.

This is needed b/c std::cin.getline() reads into a buffer inside
the istream, and that buffer was of course not visible to the
read() in get1char().

Things got tricky b/c std::cin::get() doesn't parse the ^D (EOF)
which took some digging to understand what was going on.

I also tweaked gahr's getin() to not print new lines when it wasn't
necessary, and to check cin.fail() as well as eof() out of paranoia.