Use OPENSSL_cleanse() instead of memset() in dealloc paths
Modern compilers have a model of what the memset() function does.
This is usually a great thing for optimization purposes, but it
means that it can silently remove calls if it deduces the memory
won't be written to again (such as if it's about to be deallocated)

When you are clearing memory before deallocation for security
reasons you should use something else.  Since this program already
is using OpenSSL the simplest thing is to use OPENSSL_cleanse()
which exists for precisely this purpose.